SAK-32350 Don’t fail attachments in peer review mode. (#4116)
When an assignment is in peer review student review mode there is an option for the student to provide attachments for their review, however we were attempting to check these with content-review which isn’t correct (and was failing). So this check has been disabled when in peer review mode.