KNL-1498 Reduce memory usage when handling cache invalidation arising from a realm update (#4118)
This is a compromise between creating a single large map of keys to invalidate, and calling
cache.remove() for each entry. In many cases (sites above 25 users or so), the set of potential
keys to invalidate will exceed the size of the cache. For large sites (for example 20K site members,
200 permissions), this iterates through the permissions first and does a set of cache invalidations
per permission, so the invalidation map size will be 20K. This still gains some efficiency by
allowing removeAll() to iterate through the smaller collection (the cache size).